# perf: optimize byte operations