--- a/go/mysql/server.go
+++ b/go/mysql/server.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 	"io"
 	"net"
-	"strings"
 
 	log "github.com/golang/glog"
 	"github.com/youtube/vitess/go/sqltypes"
@@ -262,7 +261,7 @@ func (l *Listener) handle(conn net.Conn, connectionID uint32) {
 		data, err := c.readEphemeralPacket()
 		if err != nil {
 			// Don't log EOF errors. They cause too much spam.
-			if !strings.HasSuffix(err.Error(), "EOF") {
+			if err != io.EOF {
 				log.Errorf("Error reading packet from client %v: %v", c.ConnectionID, err)
 			}
 			return